+++ CHANGELOG.md
@@ -11,4 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - accelerometer interrupt handler
     - charging indicator interrupt handler
     - power indicator interrupt handler
+    - backlight control
+